pyproject.toml
[tool.poetry]
name = "fast-hero"
version = "0.1.0"
description = ""
authors = ["Leonardo Melo <leo.a.s.melo@gmail.com>"]
readme = "README.md"
packages = [{include = "fast_hero"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
ignr = "^2.2"
email-validator = "^2.1.0.post1"
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.2.0"
httpx = "^0.26.0"
isort = "^5.13.2"

[tool.ruff]
line-length = 79
exclude = [".venv", 'migrations']

[tool.isort]
profile = 'black'
line_length = 79
extend_skip = ['migrations']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn fast_hero.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_hero -vv'
post_test = 'coverage html'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
